(pre-refactoring): Move UpdateRequestBuilder into a separate file
